[#DCM-146] Fix Eclipse warnings  * Fixed 'statement unnecessarily nested in else clause' warning  * Fixed 'missing @Override annotation' warning  * Fixed explicit switch fall-through warning  * Fixed 'empty block should be documented' warning  * Fixed 'unnecessary cast from int to int' warning

+++ dcm4che2/trunk/dcm4che-image/src/main/java/org/dcm4che2/image/OverlayUtils.java
@@ -211,7 +211,7 @@ public class OverlayUtils {
      */
     protected static byte[] padToFixRowByteBoundary(byte[] unpaddedData, int rows, int cols) {
         int numRowBytes = (cols+7)/8;
-        int paddedLength = (int)(rows * numRowBytes);
+        int paddedLength = rows * numRowBytes;
         if( (unpaddedData.length == paddedLength ) && (cols%8)==0 ) return unpaddedData;
         
         byte[] data = new byte[paddedLength];
